更新taskModel,修复task和用户没有关联的问题

+++ backend/controllers/taskController.js
@@ -1,8 +1,9 @@
 const asyncHandler = require("express-async-handler");
 const Task = require("../models/taskModel");
+const User = require("../models/userModel");
 
-const getTasks = asyncHandler(async (_, res) => {
-  const task = await Task.find();
+const getTasks = asyncHandler(async (req, res) => {
+  const task = await Task.find({ user: req.user.id });
   res.status(200).json(task);
 });
 
@@ -11,15 +12,25 @@ const setTask = asyncHandler(async (req, res) => {
     res.status(400);
     throw new Error("Please enter a task");
   }
-  const task = await Task.create({ text: req.body.text });
+  const task = await Task.create({ text: req.body.text, user: req.user.id });
   res.status(200).json(task);
 });
 
-const putTask = asyncHandler(async (req, res) => {
+const updateTask = asyncHandler(async (req, res) => {
   const task = await Task.findById(req.params.id);
   if (!task) {
     res.status(400);
     throw new Error("Task not found");
+  }
+  //添加用户授权校验
+  const user = await User.findById(req.user.id);
+  if (!user) {
+    res.status(401);
+    throw new Error("No such user found");
+  }
+  if (task.user.toString() != user.id) {
+    res.status(401);
+    throw new Error("User is not authorized to update");
   }
   const updatedTask = await Task.findByIdAndUpdate(req.params.id, req.body, {
     new: true,
@@ -33,8 +44,18 @@ const deleteTask = asyncHandler(async (req, res) => {
     res.status(400);
     throw new Error("Task not found");
   }
+  //添加用户授权校验
+  const user = await User.findById(req.user.id);
+  if (!user) {
+    res.status(401);
+    throw new Error("No such user found");
+  }
+  if (task.user.toString() != user.id) {
+    res.status(401);
+    throw new Error("User is not authorized to update");
+  }
   await Task.findByIdAndDelete(req.params.id);
   res.status(200).json({ id: req.params.id });
 });
 
-module.exports = { getTasks, setTask, putTask, deleteTask };
+module.exports = { getTasks, setTask, updateTask, deleteTask };
